chore(dashboard): reorganized inventory tab to be admin view only
